added lock mod
replaced SymAddrWrapper with safe AtomicSymAddr
